+++ frontend/src/common/backend.ts
@@ -87,6 +87,7 @@ export type GroupDto = {
   id: string;
   name: string;
   pictureUrl?: string;
+  memberSize?: number;
 }
 
 export type AuthorityDto = UserDto | GroupDto;
@@ -300,8 +301,8 @@ class TrustService {
 }
 
 class AuthorityService {
-  public async search(query: string): Promise<AuthorityDto[]> {
-    return axiosAuth.get<AuthorityDto[]>(`/authorities/search?query=${query}`).then(response => response.data.map(AuthorityService.fillInMissingPicture));
+  public async search(query: string, withMemberSize: boolean = false): Promise<AuthorityDto[]> {
+    return axiosAuth.get<AuthorityDto[]>(`/authorities/search?query=${query}&withMemberSize=${withMemberSize}`).then(response => response.data.map(AuthorityService.fillInMissingPicture));
   }
 
   public async listSome(authorityIds: string[]): Promise<AuthorityDto[]> {